requirements.txt
numpy==1.26.1
pillow==10.1.0
matplotlib==3.8.1
